Moved useJobsColumns and useJobsFilters under the hooks folder
